Bi-directional ManyToOne, OneToMany relationships in java code
Bi-directional relationships are flexible but *harder* to handle in
java code.

It is more natural in a relation database but Personally I will stay
away from it as far as possible. My opnion is that the database should
not dictate the application design unless needed for performance
reasons. Thus, I guess it could be used in rare cases if required.